format: Simplify csv format
Support for details such as language or word type is
just overengineering.

It's up to the user to set one root directory for each
language, simple as that.

Also the app doesn't need to know the word type. So
simplify the csv files to 2 fields: native word and
translated word.

Besides the simplification, it also makes csv files
compatible with Ankidroid.

Signed-off-by: Frederic Weisbecker <fweisbec@gmail.com>